requirements.txt
Flask==2.1.1
Flask-MySQL==1.5.2
PyICU==2.4.2
PyJWT==2.4.0
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0